pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ps_bed"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mani_skill",
    "torch",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb",
    "gymnasium",
    "mshab",
    "setuptools<82",
]

[project.optional-dependencies]
dev = ["black", "isort"]

[tool.setuptools.packages.find]
include = ["ps_bed*"]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
mshab = { path = "../mshab", editable = true }

[tool.uv]
prerelease = "allow"
override-dependencies = [
    "mplib>=0.2.1",
    "numpy>=1.26,<2.0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
